Change the policy: when decide to index the property, check only existence in the schema.  Before access mode was also checked
git-svn-id: https://svnsrv.desy.de/desy/EuXFEL/WP76/karabo/karaboFramework/branches/1.3@18512 a23ab9f9-1ea2-4934-960d-990259f9ea8c

+++ src/karabo/core/DataLogReader.cc
@@ -40,7 +40,7 @@ namespace karabo {
         
         void IndexBuilderService::buildIndexFor(const std::string& commandLineArguments) {
             //boost::mutex::scoped_lock lock(m_mutex);
-            int size = m_cache.size();
+            size_t size = m_cache.size();
             m_cache.insert(commandLineArguments);
             if (size == m_cache.size()) return;   // such a request was before
             m_svc.post(boost::bind(&IndexBuilderService::build, this, commandLineArguments));
